requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
rdkit-pypi==2022.9.5
seaborn==0.13.2
matplotlib==3.8.4
scipy==1.13.0
scikit-learn==1.4.2
umap-learn==0.5.6
